Do a better job of killing off workers. Fixes #11.
Previously, we removed the container with `singularity instance.stop`,
and assumed that the nanny would exit when its contained process
terminated. If the service is set to restart after a delay, however,
the nanny will not actually exit, but instead sleep. If, within that
delay, we restart the container with the same name, the nanny will wake
up and restart its worker processes as if nothing had happened. Now,
we explicitly kill the nanny and workers before starting new ones.